requirements.txt
beautifulsoup4==4.13.4
requests==2.32.3
python-dotenv==1.1.0
PyYAML==6.0.2
lxml==5.4.0
customtkinter==5.2.2
psycopg2==2.9.10
supabase==2.15.2
matplotlib==3.10.3